test: improve mock fixtures for e2e testing
- Reorganize fixtures: move audio folder into fixtures/ directory
- Add comprehensive fixture documentation:
  - FIXTURES_SPEC.md: authoritative specification for fixture generation
  - README.md: expanded overview with folder layout and usage
  - VERIFICATION.md: verification procedures
- Move audio generation scripts to fixtures/scripts/ for better organization
- Regenerate audio files with updated paths
- Fix test timeout: increase workflow_e2e test timeout to 20 minutes

These changes improve the structure and documentation of test fixtures,
making it easier to understand, maintain, and regenerate mock data for
offline e2e testing.
