+++ tests/fixtures/VERIFICATION.md
@@ -1,0 +1,12 @@
+# Fixture Verification
+
+This document describes how to verify that fixtures are correct
+after generation or regeneration.
+
+---
+
+## Transcript checks
+
+```bash
+ls tests/fixtures/transcripts/*.txt | wc -l
+```
